Comments and Queries -- PLEASE READ
A few notes:
I added the first sentence because we need a bit of an intro. Also, we use so many words to describe the security vendor by the end of the first paragraph that there is some confusion as to what we are talking about. I wanted to try to tie all those words together somehow. Also, we should in some way briefly define what a security vendor is--I know your readers will understand, but it is always good to be clear. I'm assuming it is a company that provides security for a company's data. Is that right? If I missed it, let me know, and I'll reword.

When I search for "Slack" is always seems to be initial capped. You can change it back to lc if that is your style. Just let me know.

What do you mean by "We want healthy redirection."?

Is "webforms" OK? Web forums?