How Become Security Expert -- Comments/Queries
The best way to really get your head around security is to attempt to exploit security vulnerabilities. "Attempt" OK?

"I'd say that the first one is to find the vulnerabilities and write the exploits for those vulnerabilities. From that, you gain an understanding of the variation of the exploits. Then you try to fix them." Made some changes to this passage. OK?

"like data validation, or data handling," These are two different things,m correct?

"You'll understand the problems of trusting who has a data," Wording OK here? Something missing? Seems a little odd.

"Because books are basically masterclasses given by the masters and experts on a chosen subject, right?" I made some changes to this sentence. OK?
Paragraph starting at 17 "And basically, we have a couple of career paths that you should that." Something missing here. Not sure what we are trying to say.

"You do a lot of paradigm shifts." I keep seeing this phrase. I'm not sure what it means. A paradigm shift refers to a fundamental change in the way things are done, yes? But you don't really do a change. You encounter changes or respond to them. You make changes. You change the way you think. But a paradigm shift is the change itself, right? If this is a normally used phrase in the industry, that's fine. But can you tell me what it means so I can be ready for it when it pops up again?

"And the next part is when you actually like the exploit. So, that's when you start to code and like the exploits yourselves. The exploits are actually tests -- they are security tests. And that's where it gets interesting! Because once you can do that, you have a way to replicate the findings. And it is not easy to get to that level. Then, you write the exploits in a testing framework because that allows you look at the ability. This is very powerful, being able to actually go and fix stuff." This paragraph is very confusing to me. I tried to clean it up. Not sure of the meaning. Not sure what it means to like the exploits. Can/Should that be made more clear? At any rate, does this paragraph read well to you? Any changes needed?

I rewrote the last two paragraphs. I don't think I changed the meaning at all -- just tightened it up. OK?

A final comment. The article seems fine. But the last two paragraphs seem to wander off topic a bit into general advice on learning. The paragraph starting at 25 might be a good place to end if you agree:
"So, that's how you do it. The faster you can do that loop, the more you become a security professional. And the more you start to understand what it takes to actually do an application in network security."