Neo4J Graph Comments/Qeries JamesRossPR
From 1st paragraph: "the candidates to commit the work for" Wording OK?
"the interaction that we have with them" Interaction with what? Candidates? Candidates, jobs, and skills? Clarify?
"and the connection with members of the team or with communities" the connections we have with members of the team ...?

2nd Paragraph: "to store, to capture and to visualize, sorry, to store and visualize" Should this be just "to store and visualize"?

"And that gives a feedback group on what actually is going on." Should it be "feedback loop"?

Read paragraph starting at 17 "And the reason for this is so you can..." Make sure it reads well and makes sense.

"We owe it to the candidates to create a great experience for them." Is this what we mean to say?

"We're also able to find, new and better, candidates and find exactly a great place to capture all the data that we already have, but when you look at it, it's already really complex, and if we don't do these, we actually get swamped. Even in basic things, when to reply? how to reply? how to note? how to put pressure? how to value? how to get feedback? So, we actually create a really powerful workflow."
I'm sorry, but I just don't know what this means. Not sure how to edit. Maybe a transcription error. If you can explain it to me, I'll rewrite this passage.