New Generation Queries/Comments JamesRossPR
"creative comments" We mean "creative commons" right? 

"and practically high works," Not sure what this means. I deleted it. But I can put it back in if we can make the meaning clear.

I had to rewrite this sentence. Is this OK? One of my comments I made to the 19-year-old I was speaking with was, "Look, get your head around these ideas. Write. Publish on sites like Medium or Amazon, for example, and get the links to those things out into the world. If you can do that you've got a job!" 

Very nice article. It kind of wanders around a bit, but it creates a positive, inspiring feeling. I had to do quite a bit of rewriting in here. I had to guess at the meaning of a lot of the sentences. I'm not sure if this is the result of transcription errors, or what. Please give it a careful read and make sure I didn't change the meaning of anything.