Update senior_cloud_engineer.md
"And the reason why we mapped this very senior person that is quite a good rate." Not sure of the meaning of this sentence: the use of the word "mapped" and the last part "that is quite a good rate."

"subtilties" Should be "subtleties" right? 

"So, they will be a running survey monkey -- someone who just tries everything." Sentence OK?

Sentence OK? "So, the idea that the system is unauthenticated, or that systems don't authorize each other, or make connections without us understanding them, from a security point of view, it's just crazy. "

I just don't understand the paragraph starting at 19: "So, Photobox Group Security provides..." Garbled language. It's not just the technical language. Transcription errors? I can't edit it. I will have to get some clarification. I'll be happy to speak with someone (Skype, Messenger, whatever you prefer) so we can talk through this. 

This sentence: "It's also a spectacular work to be used on how we consumers of those technologies, as we ourselves are moving into cloud, and our activities and understanding to scale." Again, the language is garbled.